Moved protobuf definitions out of server.rs
