+++ src/main.rs
@@ -4,9 +4,9 @@ mod DBMLib;
 mod DataReader;
 mod EdgeEval;
 mod ModelObjects;
+mod ProtobufServer;
 mod System;
 mod TransitionSystems;
-mod ProtobufServer;
 mod tests;
 
 use crate::DataReader::component_loader::{
@@ -16,10 +16,10 @@ use crate::DataReader::{parse_queries, xml_parser};
 use crate::ModelObjects::queries::Query;
 use crate::System::extract_system_rep;
 use clap::{load_yaml, App};
-use ProtobufServer::start_grpc_server_with_tokio;
 use std::env;
 use ModelObjects::component;
 use ModelObjects::queries;
+use ProtobufServer::start_grpc_server_with_tokio;
 use System::executable_query::QueryResult;
 
 #[macro_use]
